+++ client/termsetting/handlers/groupsetting.ts
@@ -257,13 +257,8 @@ function setRenderers(self: any) {
 			.classed('sjpp_grpset_addGrp_btn', true) //for integration testing
 			.style('display', 'inline-block')
 			.style('text-align', 'center')
-			.style('cursor', 'pointer')
-			.property(
-				'disabled',
-				self.opts.usecase?.regressionType == 'logistic' && self.opts.usecase?.detail == 'outcome'
-					? true
-					: self.data.groups.length >= self.maxGrpNum
-			)
+			.style('cursor', self.opts.q.mode == 'binary' ? 'default' : 'pointer')
+			.property('disabled', self.opts.q.mode == 'binary' ? true : self.data.groups.length >= self.maxGrpNum)
 			.text('Add Group')
 			.on('click', async () => {
 				newGrpNum++
@@ -577,12 +572,7 @@ function setRenderers(self: any) {
 	}
 
 	self.update = async function () {
-		self.dom.actionDiv.addGroup.property(
-			'disabled',
-			self.opts.usecase?.regressionType == 'logistic' && self.opts.usecase?.detail == 'outcome'
-				? true
-				: self.data.groups.length >= self.maxGrpNum
-		)
+		self.dom.actionDiv.addGroup.property('disabled', self.data.groups.length >= self.maxGrpNum)
 		for (const [i, grp] of self.data.groups.entries()) {
 			if (i === 0) continue
 			if (grp.currentIdx != i) {