+++ server/src/mds3.load.js
@@ -179,7 +179,7 @@ async function load_driver(q, ds) {
 
 	if (q.singleSampleMutation) {
 		if (!ds.queries.singleSampleMutation?.get) throw 'not supported on this dataset'
-		return { mlst: await ds.queries.singleSampleMutation.get(q.singleSampleMutation, q) }
+		return await ds.queries.singleSampleMutation.get(q.singleSampleMutation, q)
 	}
 	if (q.singleSampleGenomeQuantification) {
 		if (!ds.queries.singleSampleGenomeQuantification) throw 'not supported on this dataset'